change the primary key type for all entities to Long

+++ src/main/java/org/quaerense/laps/dao/EmployeeDao.java
@@ -9,7 +9,7 @@ public interface EmployeeDao {
 
     List<Employee> getEmployees();
 
-    Employee getEmployeeById(Integer id);
+    Employee getEmployeeById(Long id);
 
     Employee getEmployeeByUsername(String username);
